+++ lib/hadeploy/plugins/elastic/roles/elastic_modules/docs/generate.sh
@@ -1,0 +1,3 @@
+
+ansible-doc -M ../library/ elasticsearch_index 2>/dev/null | sed 's/[(].*elastic_modules[/]library.*[)]//' >elasticsearch_index.txt
+ansible-doc -M ../library/ elasticsearch_template 2>/dev/null | sed 's/[(].*elastic_modules[/]library.*[)]//' >elasticsearch_template.txt